+++ packages/dataviews/src/components/dataviews-layout/index.tsx
@@ -18,6 +18,7 @@ export default function DataViewsLayout( { className }: DataViewsLayoutProps ) {
 		fields,
 		getItemId,
 		getItemLevel,
+		getItemParentId,
 		hasInitiallyLoaded,
 		isLoading,
 		view,
@@ -67,6 +68,7 @@ export default function DataViewsLayout( { className }: DataViewsLayoutProps ) {
 				fields={ fields }
 				getItemId={ getItemId }
 				getItemLevel={ getItemLevel }
+				getItemParentId={ getItemParentId }
 				isLoading={ isLoading }
 				onChangeView={ onChangeView }
 				onChangeSelection={ onChangeSelection }